update tests and mappers to match cat c pattern
