+++ src/functions/uploadToRSIS/application/data-mapper/__tests__/cat-de-mapper.spec.ts
@@ -40,10 +40,9 @@ describe('mapCatDEData', () => {
 
     const expected: DataField[] =  [
       { col: 'AUTOMATIC_TEST', val: 0 },
+      { col: 'H_CODE_SAFETY_TOTAL', val: 0 },
       { col: 'REV_LEFT_TRAIL_CONT_TOTAL', val: 1 },
       { col: 'REV_LEFT_TRAIL_OBSERV_TOTAL', val: 1 },
-      { col: 'VEHICLE_CHECKS_TOTAL', val: 4 },
-      { col: 'VEHICLE_CHECKS_SERIOUS', val: 0 },
       { col: 'PRECAUTIONS_TOTAL', val: 2 },
       { col: 'CONTROL_ACC_TOTAL', val: 3 },
       { col: 'CONTROL_CLUTCH_TOTAL', val: 4 },
@@ -172,7 +171,6 @@ describe('mapCatDEData', () => {
       { col: 'ANCILLARY_CONTROLS_DANGEROUS', val: 0 },
       { col: 'EYESIGHT_COMPLETED', val: 1 },
       { col: 'REV_LEFT_TRAIL_COMPLETED', val: 1 },
-      { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
       { col: 'NORMAL_STOP_1_COMPLETED', val: 1 },
       { col: 'NORMAL_STOP_2_COMPLETED', val: 1 },
       { col: 'BUS_STOP_1_COMPLETED', val: 1 },
@@ -180,9 +178,6 @@ describe('mapCatDEData', () => {
       { col: 'ANGLED_START_COMPLETED', val: 1 },
       { col: 'UPHILL_START', val: 1 },
       { col: 'DOWN_HILL_START', val: 1 },
-      { col: 'SAFETY_QUESTIONS_FIRE_EXTINGUISHER', val: 1 },
-      { col: 'SAFETY_QUESTIONS_EMERGENCY_EXIT', val: 1 },
-      { col: 'SAFETY_QUESTIONS_FUEL_CUTOFF', val: 1 },
       { col: 'PRECAUTIONS_COMMENT', val: 'precautions fault' },
       { col: 'CONTROL_ACC_COMMENT', val: 'controls accelerator fault' },
       { col: 'CONTROL_CLUTCH_COMMENT', val: 'controls clutch fault' },
@@ -227,22 +222,18 @@ describe('mapCatDEData', () => {
       { col: 'REV_LEFT_TRAIL_OBSERV_COMMENT', val: 'reverse left observation' },
       { col: 'SHOW_ME_1_CODE', val: 'S1' },
       { col: 'SHOW_ME_1_DESCRIPTION', val: 'First Show Me Question' },
-      { col: 'SHOW_ME_2_CODE', val: 'S2' },
-      { col: 'SHOW_ME_2_DESCRIPTION', val: 'Second Show Me Question' },
-      { col: 'SHOW_ME_3_CODE', val: 'S3' },
-      { col: 'SHOW_ME_3_DESCRIPTION', val: 'Third Show Me Question' },
       { col: 'TELL_ME_1_CODE', val: 'T1' },
       { col: 'TELL_ME_1_DESCRIPTION', val: 'First Tell Me Question' },
-      { col: 'TELL_ME_2_CODE', val: 'T2' },
-      { col: 'TELL_ME_2_DESCRIPTION', val: 'Second Tell Me Question' },
       { col: 'VEHICLE_CHECKS_COMMENT', val: 'show me tell me fault' },
       { col: 'INDEPENDENT_DRIVING', val: 'Sat nav' },
+      { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
+      { col: 'VEHICLE_CHECKS_TOTAL', val: 1 },
+      { col: 'VEHICLE_CHECKS_SERIOUS', val: 1 },
       { col: 'UNCOUPLE_RECOUPLE_TOTAL', val: 0 },
       { col: 'UNCOUPLE_RECOUPLE_SERIOUS', val: 0 },
       { col: 'UNCOUPLE_RECOUPLE_DANGEROUS', val: 0 },
       { col: 'UNCOUPLE_RECOUPLE_COMPLETED', val: 0 },
     ];
-    // expect the right number of faults, with no serious or dangerous
     const result = mapCatD1EData(fullyPopulated);
     expect(result).toEqual(expected);
   });
@@ -250,14 +241,12 @@ describe('mapCatDEData', () => {
   it('Should map a fully populated regular test result (every possible serious fault)', () => {
     const fullyPopulated = getFullyPopulatedSeriousFaults(getMinimalInput(TestCategory.DE));
     const expected = getCatDEFullyPopulatedSeriousDataFields();
-    // expect all serious, no faults or dangerous
     expect(mapCatD1EData(fullyPopulated)).toEqual(expected);
   });
 
   it('Should map a fully populated regular test result (every possible dangerous fault)', () => {
     const fullyPopulated = getFullyPopulatedDangerousFaults(getMinimalInput(TestCategory.DE));
     const expected = getCatDEFullyPopulatedDangerousDataFields();
-    // expect all dangerous, no faults or serious
     expect(mapCatD1EData(fullyPopulated)).toEqual(expected);
   });
 });
