+++ src/functions/uploadToRSIS/application/data-mapper/__tests__/helpers/cat-de/data-fields/fully-populated-data-fields.ts
@@ -3,10 +3,9 @@ import { DataField } from '../../../../../../domain/mi-export-data';
 export function getCatDEFullyPopulatedSeriousDataFields(): DataField[] {
   return [
     { col: 'AUTOMATIC_TEST', val: 0 },
+    { col: 'H_CODE_SAFETY_TOTAL', val: 0 },
     { col: 'REV_LEFT_TRAIL_CONT_TOTAL', val: 0 },
     { col: 'REV_LEFT_TRAIL_OBSERV_TOTAL', val: 0 },
-    { col: 'VEHICLE_CHECKS_TOTAL', val: 4 },
-    { col: 'VEHICLE_CHECKS_SERIOUS', val: 1 },
     { col: 'PRECAUTIONS_TOTAL', val: 0 },
     { col: 'CONTROL_ACC_TOTAL', val: 0 },
     { col: 'CONTROL_CLUTCH_TOTAL', val: 0 },
@@ -135,7 +134,6 @@ export function getCatDEFullyPopulatedSeriousDataFields(): DataField[] {
     { col: 'ANCILLARY_CONTROLS_DANGEROUS', val: 0 },
     { col: 'EYESIGHT_COMPLETED', val: 1 },
     { col: 'REV_LEFT_TRAIL_COMPLETED', val: 1 },
-    { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
     { col: 'NORMAL_STOP_1_COMPLETED', val: 1 },
     { col: 'NORMAL_STOP_2_COMPLETED', val: 1 },
     { col: 'BUS_STOP_1_COMPLETED', val: 1 },
@@ -143,9 +141,6 @@ export function getCatDEFullyPopulatedSeriousDataFields(): DataField[] {
     { col: 'ANGLED_START_COMPLETED', val: 1 },
     { col: 'UPHILL_START', val: 1 },
     { col: 'DOWN_HILL_START', val: 1 },
-    { col: 'SAFETY_QUESTIONS_FIRE_EXTINGUISHER', val: 1 },
-    { col: 'SAFETY_QUESTIONS_EMERGENCY_EXIT', val: 1 },
-    { col: 'SAFETY_QUESTIONS_FUEL_CUTOFF', val: 1 },
     { col: 'EYESIGHT_COMMENT', val: 'eyesight serious' },
     { col: 'PRECAUTIONS_COMMENT', val: 'precautions serious' },
     { col: 'CONTROL_ACC_COMMENT', val: 'controls accelerator serious' },
@@ -191,16 +186,13 @@ export function getCatDEFullyPopulatedSeriousDataFields(): DataField[] {
     { col: 'REV_LEFT_TRAIL_OBSERV_COMMENT', val: 'reverse left observation' },
     { col: 'SHOW_ME_1_CODE', val: 'S1' },
     { col: 'SHOW_ME_1_DESCRIPTION', val: 'First Show Me Question' },
-    { col: 'SHOW_ME_2_CODE', val: 'S2' },
-    { col: 'SHOW_ME_2_DESCRIPTION', val: 'Second Show Me Question' },
-    { col: 'SHOW_ME_3_CODE', val: 'S3' },
-    { col: 'SHOW_ME_3_DESCRIPTION', val: 'Third Show Me Question' },
     { col: 'TELL_ME_1_CODE', val: 'T1' },
     { col: 'TELL_ME_1_DESCRIPTION', val: 'First Tell Me Question' },
-    { col: 'TELL_ME_2_CODE', val: 'T2' },
-    { col: 'TELL_ME_2_DESCRIPTION', val: 'Second Tell Me Question' },
     { col: 'VEHICLE_CHECKS_COMMENT', val: 'show me tell me serious' },
     { col: 'INDEPENDENT_DRIVING', val: 'Traffic signs' },
+    { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
+    { col: 'VEHICLE_CHECKS_TOTAL', val: 1 },
+    { col: 'VEHICLE_CHECKS_SERIOUS', val: 1 },
     { col: 'UNCOUPLE_RECOUPLE_TOTAL', val: 0 },
     { col: 'UNCOUPLE_RECOUPLE_SERIOUS', val: 0 },
     { col: 'UNCOUPLE_RECOUPLE_DANGEROUS', val: 0 },
@@ -211,10 +203,9 @@ export function getCatDEFullyPopulatedSeriousDataFields(): DataField[] {
 export function getCatDEFullyPopulatedDangerousDataFields(): DataField[] {
   return [
     { col: 'AUTOMATIC_TEST', val: 0 },
+    { col: 'H_CODE_SAFETY_TOTAL', val: 0 },
     { col: 'REV_LEFT_TRAIL_CONT_TOTAL', val: 0 },
     { col: 'REV_LEFT_TRAIL_OBSERV_TOTAL', val: 0 },
-    { col: 'VEHICLE_CHECKS_TOTAL', val: 4 },
-    { col: 'VEHICLE_CHECKS_SERIOUS', val: 0 },
     { col: 'PRECAUTIONS_TOTAL', val: 0 },
     { col: 'CONTROL_ACC_TOTAL', val: 0 },
     { col: 'CONTROL_CLUTCH_TOTAL', val: 0 },
@@ -343,7 +334,6 @@ export function getCatDEFullyPopulatedDangerousDataFields(): DataField[] {
     { col: 'ANCILLARY_CONTROLS_DANGEROUS', val: 1 },
     { col: 'EYESIGHT_COMPLETED', val: 1 },
     { col: 'REV_LEFT_TRAIL_COMPLETED', val: 1 },
-    { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
     { col: 'NORMAL_STOP_1_COMPLETED', val: 1 },
     { col: 'NORMAL_STOP_2_COMPLETED', val: 1 },
     { col: 'BUS_STOP_1_COMPLETED', val: 1 },
@@ -351,9 +341,6 @@ export function getCatDEFullyPopulatedDangerousDataFields(): DataField[] {
     { col: 'ANGLED_START_COMPLETED', val: 1 },
     { col: 'UPHILL_START', val: 1 },
     { col: 'DOWN_HILL_START', val: 1 },
-    { col: 'SAFETY_QUESTIONS_FIRE_EXTINGUISHER', val: 1 },
-    { col: 'SAFETY_QUESTIONS_EMERGENCY_EXIT', val: 1 },
-    { col: 'SAFETY_QUESTIONS_FUEL_CUTOFF', val: 1 },
     { col: 'PRECAUTIONS_COMMENT', val: 'precautions dangerous' },
     { col: 'CONTROL_ACC_COMMENT', val: 'controls accelerator dangerous' },
     { col: 'CONTROL_CLUTCH_COMMENT', val: 'controls clutch dangerous' },
@@ -398,15 +385,12 @@ export function getCatDEFullyPopulatedDangerousDataFields(): DataField[] {
     { col: 'REV_LEFT_TRAIL_OBSERV_COMMENT', val: 'reverse left observation' },
     { col: 'SHOW_ME_1_CODE', val: 'S1' },
     { col: 'SHOW_ME_1_DESCRIPTION', val: 'First Show Me Question' },
-    { col: 'SHOW_ME_2_CODE', val: 'S2' },
-    { col: 'SHOW_ME_2_DESCRIPTION', val: 'Second Show Me Question' },
-    { col: 'SHOW_ME_3_CODE', val: 'S3' },
-    { col: 'SHOW_ME_3_DESCRIPTION', val: 'Third Show Me Question' },
     { col: 'TELL_ME_1_CODE', val: 'T1' },
     { col: 'TELL_ME_1_DESCRIPTION', val: 'First Tell Me Question' },
-    { col: 'TELL_ME_2_CODE', val: 'T2' },
-    { col: 'TELL_ME_2_DESCRIPTION', val: 'Second Tell Me Question' },
     { col: 'INDEPENDENT_DRIVING', val: 'N/A' },
+    { col: 'VEHICLE_CHECKS_COMPLETED', val: 1 },
+    { col: 'VEHICLE_CHECKS_TOTAL', val: 0 },
+    { col: 'VEHICLE_CHECKS_SERIOUS', val: 0 },
     { col: 'UNCOUPLE_RECOUPLE_TOTAL', val: 0 },
     { col: 'UNCOUPLE_RECOUPLE_SERIOUS', val: 0 },
     { col: 'UNCOUPLE_RECOUPLE_DANGEROUS', val: 0 },
